llm-gemini-chatbot/backend/requirements.txt
fastapi 
uvicorn 
langchain 
langserve 
streamlit 
google-genai 
python-dotenv
langchain-google-genai
requests
